requirements.txt
Django>=5.1.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
drf-yasg>=1.21.7
django-filter>=23.5 